회원기능 만들기: OAuth + session방식 사용하기

+++ forum/pages/api/list/[id].ts
@@ -1,14 +1,19 @@
 import { DB } from "@/components/constants";
 import {
   deleteOneDocument,
+  findOneDocument,
   updateOneDocument,
 } from "@/components/util/database";
 import {
   response200,
   response400,
+  response401,
   response405,
 } from "@/components/util/server";
 import { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
+import { ObjectId } from "mongodb";
 
 export default async function handler(
   req: NextApiRequest,
@@ -27,26 +32,35 @@ export default async function handler(
         id,
         putData
       );
-      res.redirect(302, "/list");
-      break;
+      return res.redirect(302, "/list");
 
     case "DELETE":
+      const session = await getServerSession(req, res, authOptions);
+
+      if (!session || !session.user || !session.user.email)
+        return response401(res);
+
+      const doc = await findOneDocument(
+        DB.APPLE_FORUM.NAME,
+        DB.APPLE_FORUM.POST,
+        {
+          _id: new ObjectId(id),
+        }
+      );
+
+      if (doc?.author !== session?.user?.email) return response401(res);
+
       const result = await deleteOneDocument(
         DB.APPLE_FORUM.NAME,
         DB.APPLE_FORUM.POST,
         id
       );
 
-      if (result.deletedCount < 1) {
-        response400(res);
-        break;
-      }
+      if (result.deletedCount < 1) return response400(res);
 
-      response200(res, "삭제완");
-      break;
+      return response200(res, "삭제완");
 
     default:
-      response405(res, "PUT", "DELETE");
-      break;
+      return response405(res, "PUT", "DELETE");
   }
 }
